processor/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
python-docx==1.1.0
pdfminer.six==20231228
python-pptx==0.6.23
pytesseract==0.3.10
paddleocr==2.7.3
Pillow==10.1.0
opencv-python<=4.6.0.66
numpy==1.26.4
setuptools>=68
wheel>=0.41
python-magic==0.4.27
pydantic==2.5.0
pydantic-settings==2.1.0
pix2tex[api]
mss>=9.0.0
pyperclip>=1.8.0
PyMuPDF==1.24.9
python-dotenv==1.0.1